refactor(ssh): refactor packages and add sshInfoCreateRequestDto
