+++ triton_dashboard/src/main/java/com/triton/msa/triton_dashboard/project/entity/Project.java
@@ -33,4 +33,8 @@ public class Project {
 
     @OneToMany(mappedBy = "project", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<PrivateData> privateData = new ArrayList<>();
+
+    public void updateSshInfo(SshInfo sshInfo) {
+        this.sshInfo = sshInfo;
+    }
 }
