+++ triton_dashboard/src/main/java/com/triton/msa/triton_dashboard/ssh/controller/SshApiController.java
@@ -2,7 +2,7 @@ package com.triton.msa.triton_dashboard.ssh.controller;
 
 import com.triton.msa.triton_dashboard.project.entity.Project;
 import com.triton.msa.triton_dashboard.project.service.ProjectService;
-import com.triton.msa.triton_dashboard.ssh.service.SshService;
+import com.triton.msa.triton_dashboard.ssh.client.SshServiceClient;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -14,10 +14,10 @@ import java.util.Map;
 @RestController
 @RequestMapping("/api/ssh")
 public class SshApiController {
-    private final SshService sshService;
+    private final SshServiceClient sshService;
     private final ProjectService projectService;
 
-    public SshApiController(SshService sshService, ProjectService projectService) {
+    public SshApiController(SshServiceClient sshService, ProjectService projectService) {
         this.sshService = sshService;
         this.projectService = projectService;
     }
