+++ triton_dashboard/src/main/java/com/triton/msa/triton_dashboard/ssh/dto/SshInfoCreateRequestDto.java
@@ -1,0 +1,10 @@
+package com.triton.msa.triton_dashboard.ssh.dto;
+
+import org.springframework.web.multipart.MultipartFile;
+
+public record SshInfoCreateRequestDto(
+        String sshIpAddress,
+        String username,
+        MultipartFile pemFile
+) {
+}
